interpolation resolution setting

+++ src/DataHandler.hpp
@@ -32,7 +32,7 @@ public:
 
     uint8_t x_axis_select = 0;
     bool interpolation = false;
-    uint8_t ip_values_per_step = 10;
+    int ip_values_per_step = 10;
 
 private:
     uint32_t current_absolute_sample = 0;
